Fix C4293 and C4307 in SendProxy_CropFlagsToPlayerFlagBitsLength (#867)

+++ src/game/server/player.cpp
@@ -8103,7 +8103,7 @@ void CMovementSpeedMod::InputSpeedMod(inputdata_t &data)
 
 void SendProxy_CropFlagsToPlayerFlagBitsLength( const SendProp* pProp, const void* pStruct, const void* pVarData, DVariant* pOut, int iElement, int objectID )
 {
-	int mask = ( 1 << PLAYER_FLAG_BITS ) - 1;
+	int mask = static_cast< unsigned >( 1 << PLAYER_FLAG_BITS ) - 1;
 	int data = *( int* )pVarData;
 
 	pOut->m_Int = ( data & mask );
